split lexer into source and header
