+++ src/lexer.cpp
@@ -1,51 +1,15 @@
 #include <string>
-#include <vector>
 #include <tuple>
 #include <algorithm>
 #include <cassert>
-#include <experimental/string_view>
 #include <fmt/format.h>
 #include <fmt/format.cc>
 #include "cpp/contracts.hpp"
 #include "cpp/optional.hpp"
+#include "lexer.hpp"
 
-using std::experimental::string_view;
-
-enum class TokenType
+namespace
 {
-  // Symbols
-  Plus,
-  Minus,
-  Times,
-  Divide,
-  Assign,
-  LeftParen,
-  RightParen,
-  LeftBraces,
-  RightBraces,
-  LeftCurlyBraces,
-  RightCurlyBraces,
-  StringMark,
-  CharMark,
-
-  // Constants
-  IntegerConstant,
-  FloatConstant,
-
-  // Qualified ids
-  Identifier,
-
-  // Reserved names
-  If,
-  Else,
-  For,
-  While,
-  Do,
-
-  // Types
-  IntType,
-  FloatType,
-};
 
 const auto TOKEN_SYMBOLS = std::vector<std::pair<TokenType, string_view>>
 {
@@ -65,6 +29,9 @@ const auto TOKEN_SYMBOLS = std::vector<std::pair<TokenType, string_view>>
   {TokenType::RightCurlyBraces, "}"},
   {TokenType::StringMark,       "\""},
   {TokenType::CharMark,         "'"},
+  {TokenType::Comma,            ","},
+  {TokenType::Colon,            ":"},
+  {TokenType::Semicolon,        ";"},
 };
 
 const auto TOKEN_RESERVED_NAMES = std::vector<std::pair<TokenType, string_view>>
@@ -78,7 +45,6 @@ const auto TOKEN_RESERVED_NAMES = std::vector<std::pair<TokenType, string_view>>
   {TokenType::FloatType, "float"},
 };
 
-
 // Whites are used to tell apart from different tokens.
 constexpr auto is_white(char c) -> bool
 {
@@ -90,7 +56,8 @@ constexpr auto is_white(char c) -> bool
          c == '}'  ||
          c == ','  ||
          c == ':'  ||
-         c == '?';
+         c == '?'  ||
+         c == ';';
 }
 
 constexpr auto is_space(char c) -> bool
@@ -141,22 +108,6 @@ auto find_token_between_whites(const char* begin, const char* end) -> optional<s
   return nullopt;
 }
 
-
-struct TokenData
-{
-  TokenType type;
-  string_view data;
-
-  constexpr explicit TokenData(TokenType type, const char* begin, const char* end) noexcept :
-    type(type),
-    data(begin, static_cast<size_t>(end - begin))
-  {}
-
-  constexpr explicit TokenData(TokenType type, string_view source) noexcept :
-    type(type),
-    data(source.begin(), static_cast<size_t>(source.end() - source.begin()))
-  {}
-};
 
 struct LexerContext
 {
@@ -316,142 +267,11 @@ auto lexer_parse_identifier(LexerContext& lexer, const char* begin, const char* 
   return token.end();
 }
 
-namespace test
+
+} // namespace
+
+auto lexer_parse(const std::string& data) -> std::vector<TokenData>
 {
-
-struct LexerState
-{
-  string_view input;
-  LexerContext context;
-  const char* it;
-
-  explicit LexerState(string_view input) noexcept :
-    input{input}, context{}, it{input.begin()}
-  {}
-
-  template <typename F>
-  auto operator() (F&& f) noexcept -> const TokenData&
-  {
-    it = std::forward<F>(f)(context, it, input.end());
-    return context.tokens.back();
-  }
-};
-
-void test_lexer_parse_constant()
-{
-  fmt::print(stderr, "Testing lexer_parse_constant ... ");
-
-  // IntegerConstants.
-  {
-    LexerState lexer{"42 0x22fx"};
-
-    auto t1 = lexer(lexer_parse_constant);
-
-    assert(t1.type == TokenType::IntegerConstant);
-    assert(t1.data == "42");
-
-    // Should fail
-    auto t2 = lexer(lexer_parse_constant);
-
-    assert(lexer.context.tokens.size() == 1);
-    assert(t2.type == TokenType::IntegerConstant);
-    assert(t2.data == "42");
-  }
-
-  // FloatConstants.
-  {
-    LexerState lexer{"123.f 0.2 1. (1.f) 0.ff .0f"};
-
-    {
-      auto t = lexer(lexer_parse_constant);
-      assert(t.type == TokenType::FloatConstant);
-      assert(t.data == "123.f");
-    }
-
-    {
-      auto t = lexer(lexer_parse_constant);
-      assert(t.type == TokenType::FloatConstant);
-      assert(t.data == "0.2");
-    }
-
-    {
-      auto t = lexer(lexer_parse_constant);
-      assert(t.type == TokenType::FloatConstant);
-      assert(t.data == "1.");
-    }
-
-    {
-      auto t = lexer(lexer_parse_constant);
-      assert(t.type == TokenType::FloatConstant);
-      assert(t.data == "1.f");
-    }
-
-    {
-      // Should fail; nothing gets pushed into context.
-      auto t = lexer(lexer_parse_constant);
-      assert(t.type == TokenType::FloatConstant);
-      assert(t.data == "1.f");
-
-      // Should also skip ill-formed constant.
-    }
-
-    {
-      auto t= lexer(lexer_parse_constant);
-      assert(t.type == TokenType::FloatConstant);
-      assert(t.data == ".0f");
-    }
-  }
-
-  fmt::print(stderr, "Success.\n");
-}
-
-void test_lexer_parse_identifier()
-{
-  fmt::print(stderr, "Testing lexer_parse_identifier ... ");
-
-  LexerState lexer{"abc42() _KeepMoving_Forward 42fail{}success_1 if"};
-
-  {
-    auto t = lexer(lexer_parse_identifier);
-    assert(t.type == TokenType::Identifier);
-    assert(t.data == "abc42");
-  }
-
-  {
-    auto t = lexer(lexer_parse_identifier);
-    assert(t.type == TokenType::Identifier);
-    assert(t.data == "_KeepMoving_Forward");
-  }
-
-  {
-    // Should fail; nothing gets pushed into context.
-    auto t = lexer(lexer_parse_identifier);
-    assert(t.type == TokenType::Identifier);
-    assert(t.data == "_KeepMoving_Forward");
-
-    // Should also skip ill-formed identifier.
-  }
-
-  {
-    auto t = lexer(lexer_parse_identifier);
-    assert(t.type == TokenType::Identifier);
-    assert(t.data == "success_1");
-  }
-
-  {
-    auto t = lexer(lexer_parse_identifier);
-    assert(t.type == TokenType::If);
-    assert(t.data == "if");
-  }
-
-  fmt::print(stderr, "Success.\n");
-}
-
-}
-
-int main()
-{
-  test::test_lexer_parse_constant();
-  test::test_lexer_parse_identifier();
+  return {};
 }
